2.2.2

+++ 2.2.2. spring_mvc/src/main/java/model/Car.java
@@ -1,23 +1,22 @@
 package model;
 
 public class Car {
-    private Long id;
-    //@Column(name = "name")
+    private int id;
     private String model;
-    //@Column(name = "lastName")
     private String color;
 
     public Car(){
 
     }
 
-    public Car(Long id, String model, String color) {
+    public Car(int id, String model, String color) {
         this.id = id;
         this.model = model;
         this.color = color;
     }
 
-    public Long getId() {
+
+    public int getId() {
         return id;
     }
 
@@ -29,7 +28,7 @@ public class Car {
         return color;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
